+++ src/App/Documentation/core/Page-layout/index.js
@@ -1,0 +1,76 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import { Property, ComponentPreview, DocContainer } from "#";
+
+const FullWidth = () => (
+    <>
+        <h2 id="full-width">Full width page</h2>
+        <p>
+            To use a full width layout use a combination of grid and the <Property value=".main-content" /> class. <Property value=".main-content" /> has a white background and some padding.
+            For a full width layout use <Property value=".border-left" /> to clearly seperate the sidebar and the main content. See <Link to="/docs/utilities/borders">borders</Link> docs for more information.
+        </p>
+        <ComponentPreview language="html" codeFigure>
+            <div className="container-fluid">
+                <div className="row">
+                    <header>
+                        Topbar component...
+                    </header>
+                </div>
+                <div className="row">
+                    <div className="col-auto">
+                        Sidebar component...
+                    </div>
+                    <main className="col main-content border-left">
+                        Main content things...
+                    </main>
+                </div>
+            </div>
+        </ComponentPreview>
+    </>
+);
+
+const CenteredContent = () => (
+    <>
+        <h2 id="centered-content">Centered content</h2>
+        <p>
+            For a centered layout use <Property value=".container"/>, it has built in breakpoints and will switch to full-width if the screen size is small
+            enough (see <Link to="/docs/core/grid">grid</Link>). When the content is centered it will require <Link to="/docs/utilities/borders">borders</Link> on
+            the right and left side. To make your topbar fill the page horizontally you will have to wrap it in a <Property value=".row" /> and <Property value=".container-fluid" />.
+        </p>
+        <ComponentPreview language="html" codeFigure>
+            <div className="container-fluid">
+                <div className="row">
+                    <header>
+                        Topbar component...
+                    </header>
+                </div>
+            </div>
+            <div className="container">
+                <div className="row">
+                    <div className="col-auto">
+                        Sidebar component...
+                    </div>
+                    <main className="col main-content border-left border-right">
+                        Main content things...
+                    </main>
+                </div>
+            </div>
+        </ComponentPreview>
+    </>
+);
+
+const PageLayout = () => (
+    <DocContainer docToc >
+        <p className="lead">
+            Use grid with our <Property value=".main-content" /> class to use a variety of page layouts. As we have a mobile first approach we only use 1rem as base padding for
+            our <Property value=".main-content"/> class. Use the <Link to="/docs/utilities/spacing">spacing</Link> classes to customize padding to fit your needs.
+        </p>
+        <FullWidth />
+        <CenteredContent />
+    </DocContainer>
+);
+
+export default PageLayout;
+
+/* For testing */
+export { FullWidth, CenteredContent };