+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ### Added
 
+- Added Page Layout section /core
+- Added .main-content class
+- Added Border section to /utilities
+- Added .border class along with more variants
+
 ### Changed
 
 ### Fixed